net/ixgbe: add VF promiscuous mode support when PF as host
There is need to PF host promiscuous mode enable. For ixgbe,
in order to support VF vlan promiscuous or unicast promiscuous,
we need to set PF host register PFVML2FLT of bit UPE and VPE.
It also align to ixgbe kernel code version 5.5.3. And also
it need to update to the latest API version in order for
negotiation between pf and vf, then new API message can be
supported.

Signed-off-by: Wei Zhao <wei.zhao1@intel.com>
Signed-off-by: 0-day Robot <robot_email>