net/ice/base: allow package copy to be used after resets
For components that make a copy of an external pipeline package file (i.e.
the Linux and FreeBSD drivers), save the size of the package file along
with the copy so that both can be used when calling ice_init_pkg() after a
CORER/GLOBR reset.  Also, do not free the copy of the package file in
ice_init_pkg() since it is needed afterward for subsequent resets.

Signed-off-by: Bruce Allan <bruce.w.allan@intel.com>
Signed-off-by: Paul M Stillwell Jr <paul.m.stillwell.jr@intel.com>
Signed-off-by: Qi Zhang <qi.z.zhang@intel.com>
Signed-off-by: 0-day Robot <robot_email>