net/mlx5: replace IPC socket with EAL API
Socket API is used for IPC in order for secondary process to acquire Verb
command file descriptor. The FD is used to remap UAR address. The new
multi-process APIs (rte_mp) in EAL are newly introduced. mlx5_socket.c is
replaced with mlx5_mp.c, which uses the new APIs.

As it is PMD global infrastructure, only one IPC channel is established.
All the IPC message types may have port_id in the message if there is need
to reference a specific device.

Signed-off-by: Yongseok Koh <yskoh@mellanox.com>
Acked-by: Shahaf Shuler <shahafs@mellanox.com>
Signed-off-by: 0-day Robot <robot_email>