net/mlx4: add external allocator for Verbs object
To support secondary process, the memory allocated by library such as
completion rings (CQ) and buffer rings (WQ) must be manageable by EAL, in
order to share it with secondary processes. With new changes in rdma-core
and kernel driver, it is possible to provide an external allocator to the
library layer for this purpose. All such resources will now be allocated
within DPDK framework.

Signed-off-by: Yongseok Koh <yskoh@mellanox.com>
Acked-by: Shahaf Shuler <shahafs@mellanox.com>
Signed-off-by: 0-day Robot <robot_email>